add integer only output feature use vector to store decodings

+++ decoding/catCSV.sh
@@ -10,11 +10,14 @@
 #catAllCSV=1 enables cat of entire csv files of all dates generated as output into single csv file, catAllCSV=0 disables it
 catAllCSV=1
 
-# =1 enables collecting cat csv in separate file for each station
+# =1 enables collecting cat csv in separate file for each station. to separate, the first column of csv should be station code
 separateFilePerStation=1
 
 #enable sorting according to station code
 sort=1
+
+#enables field header in csv files
+enableHeader=1
 
 outputDir="../Input-Output/catCSV"
 initialInputDir="../Input-Output/CSV"
@@ -22,8 +25,12 @@ outFileName="DEMS.csv"
 
 rm -Rf "$outputDir" && mkdir -p "$outputDir"
 
+csvHeader=""
+
 subdirs=$(ls "$initialInputDir")
 
+tmpFile=$(date +%d%m%Y)
+    
 for currDate in $subdirs ;do
 
     mkdir -p "$outputDir/$currDate"
@@ -32,40 +39,59 @@ for currDate in $subdirs ;do
         filename="$initialInputDir/$currDate/$x"
 
         lines=$(wc -l < $filename)
-        if [ $lines ] ;then
+        if [ $lines -ne 0 ] ;then #file has some content
+            if [ -z "$csvHeader" ] ;then #read csv header for first time only
+                csvHeader=$(head -n 1 $filename)
+            fi
+        
             tail -n $((lines-1)) < $filename >> "$outputDir/$currDate/$outFileName"
         fi
         i=$((i+1))
     done
     
-    tmpFile=$(date +%d%m%Y)
-    
-    #sort according to station code
     if [ $sort -eq 1 -a $catAllCSV -eq 0 ] ;then
+        #sort according to station code
         cat "$outputDir/$currDate/$outFileName" | tr "," " " | sort -n -k1 | sed s/"\" \""/"\",\""/g >> "$outputDir/$currDate/$tmpFile"
-        mv "$outputDir/$currDate/$tmpFile" "$outputDir/$currDate/$outFileName"
+        
+        if [ $enableHeader -eq 1 ] ;then
+            echo "$csvHeader" > "$outputDir/$currDate/$outFileName"
+            cat "$outputDir/$currDate/$tmpFile" >> "$outputDir/$currDate/$outFileName"
+        else
+            mv "$outputDir/$currDate/$tmpFile" "$outputDir/$currDate/$outFileName"
+        fi
     fi
 done
 
 if [ $catAllCSV -eq 1 ] ;then
     catCSVsubdirs=$( ls "$outputDir" )
     for currDate in $catCSVsubdirs ; do
-        cat $outputDir/$currDate/$outFileName >> $outputDir/$outFileName
+        if [ -e $outputDir/$currDate/$outFileName ] ;then #there exists cat file for currdate
+            cat $outputDir/$currDate/$outFileName >> $outputDir/$outFileName
+        fi
         rm -R $outputDir/$currDate
     done
+    
+    if [ $sort -eq 1 -a $separateFilePerStation -eq 0 ] ;then
+        cat "$outputDir/$outFileName" | tr "," " " | sort -n -k1 | sed s/"\" \""/"\",\""/g >> "$outputDir/$tmpFile"
+        mv "$outputDir/$tmpFile" "$outputDir/$outFileName"
+    fi
+    if [ $enableHeader -eq 1 -a $separateFilePerStation -eq 0 ] ;then
+        echo -e "$csvHeader\n$(cat $outputDir/$outFileName)" > $outputDir/$outFileName
+    fi
 fi
 
-if [ $sort -eq 1 -a $catAllCSV -eq 1 -a $separateFilePerStation -eq 0 ] ;then
-    tmpFile=$(date +%d%m%Y)
-    cat "$outputDir/$outFileName" | tr "," " " | sort -n -k1 | sed s/"\" \""/"\",\""/g >> "$outputDir/$tmpFile"
-    mv "$outputDir/$tmpFile" "$outputDir/$outFileName"
-fi
-
-if [ $separateFilePerStation -eq 1 ] ;then
+if [ $separateFilePerStation -eq 1 -a $catAllCSV -eq 1 ] ;then
     while IFS='' read -r line || [[ -n "$line" ]]; do
         stnCode=$(echo "$line" | cut -d, -f1 | sed s/"\""/""/g )
         echo "$line" >> "$outputDir/$stnCode.csv"
     done < "$outputDir/$outFileName"
     
     rm "$outputDir/$outFileName"
+    
+    if [ $enableHeader -eq 1 ] ;then
+        files=$(ls $outputDir | grep ".csv")
+        for x in $files ;do
+            echo -e "$csvHeader\n$(cat $outputDir/$x)" > $outputDir/$x
+        done
+    fi
 fi
